# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Developnow Test-Api Swagger Documentation",
        "description": "Developnow Api Swagger documentation description",
        "contact": {
            "email": "veshraj.joshi1@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "Developnow Test API Server"
        }
    ],
    "paths": {
...
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get list of products",
                "description": "Returns list of products",
                "operationId": "products-list",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Current page"
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "description": "Page Size"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SwaggerProductResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Forbidden"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Store new products",
                "description": "Returns the created object of products",
                "operationId": "products-create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SwaggerProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "status",
                                            "type": "string"
                                        },
                                        "message": {
                                            "description": "Item Created successfully",
                                            "type": "string"
                                        },
                                        "payload": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SwaggerProduct"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Forbidden"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/products/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get products information",
                "description": "Returns products data",
                "operationId": "products-details",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "products id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "status",
                                            "type": "string"
                                        },
                                        "message": {
                                            "description": "Item Created successfully",
                                            "type": "string"
                                        },
                                        "payload": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SwaggerProduct"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Forbidden"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "Update existing products",
                "description": "Returns updated products data",
                "operationId": "products-update",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SwaggerProductUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "status",
                                            "type": "string"
                                        },
                                        "message": {
                                            "description": "Item Created successfully",
                                            "type": "string"
                                        },
                                        "payload": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SwaggerProduct"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Forbidden"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete existing products",
                "description": "Deletes a record and returns no content",
                "operationId": "products-delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "products id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Forbidden"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Forbidden": {
                "title": "Forbidden",
                "description": "Forbidden model",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "status",
                        "type": "string",
                        "format": "string",
                        "example": "err"
                    },
                    "message": {
                        "title": "message",
                        "description": "message",
                        "type": "string",
                        "format": "string",
                        "example": "Either this content doesnot exists or you do not have enough priviledges"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Forbidden"
                }
            },
...
            "NotFound": {
                "title": "NotFound",
                "description": "NotFound model",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "status",
                        "type": "string",
                        "format": "string",
                        "example": "err"
                    },
                    "message": {
                        "title": "message",
                        "description": "message",
                        "type": "string",
                        "format": "string",
                        "example": "404 Page not found"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "NotFound"
                }
            },
...
            "Unauthenticated": {
                "title": "Unauthenticated",
                "description": "Unauthenticated model",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "status",
                        "type": "string",
                        "format": "string",
                        "example": "err"
                    },
                    "message": {
                        "title": "message",
                        "description": "message",
                        "type": "string",
                        "format": "string",
                        "example": "You are not logged in. Please login first."
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Unauthenticated"
                }
            },
...
            "SwaggerProduct": {
                "title": "Product",
                "description": "Product model",
                "xml": {
                    "name": "Product"
                },
                "allOf": [
                    {
                        "$ref": "#/components/schemas/BaseModel"
                    },
                    {
                        "properties": {
                            "name": {
                                "title": "name",
                                "description": "name",
                                "type": "string",
                                "format": "string",
                                "example": ""
                            },
                            "price": {
                                "title": "price",
                                "description": "price",
                                "type": "string",
                                "format": "string",
                                "example": ""
                            },
                            "category_id": {
                                "title": "category_id",
                                "description": "category_id",
                                "type": "integer",
                                "format": "string",
                                "example": 1
                            },
                            "image": {
                                "title": "image",
                                "description": "image",
                                "type": "string",
                                "format": "string",
                                "example": "image_url"
                            },
                            "inventory_stock": {
                                "title": "inventory_stock",
                                "description": "inventory_stock",
                                "type": "integer",
                                "format": "string",
                                "example": 1
                            },
                            "category": {
                                "$ref": "#/components/schemas/SwaggerProductCategory"
                            }
                        }
                    }
                ]
            },
...
            "SwaggerProductRequest": {
                "title": " Product  request",
                "description": "  Product  request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "name",
                        "type": "string",
                        "format": "string",
                        "example": ""
                    },
                    "price": {
                        "title": "price",
                        "description": "price",
                        "type": "string",
                        "format": "string",
                        "example": ""
                    },
                    "category_id": {
                        "title": "category_id",
                        "description": "category_id",
                        "type": "integer",
                        "format": "string",
                        "example": 1
                    },
                    "image": {
                        "title": "image",
                        "description": "image",
                        "type": "string",
                        "format": "string",
                        "example": "image_url"
                    },
                    "inventory_stock": {
                        "title": "inventory_stock",
                        "description": "inventory_stock",
                        "type": "integer",
                        "format": "string",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "SwaggerProductUpdateRequest": {
                "title": "Update Product  request",
                "description": "Update  Product  request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "string",
                        "format": "string",
                        "example": ""
                    },
                    "name": {
                        "title": "name",
                        "description": "name",
                        "type": "string",
                        "format": "string",
                        "example": ""
                    },
                    "price": {
                        "title": "price",
                        "description": "price",
                        "type": "string",
                        "format": "string",
                        "example": ""
                    },
                    "category_id": {
                        "title": "category_id",
                        "description": "category_id",
                        "type": "integer",
                        "format": "string",
                        "example": 1
                    },
                    "image": {
                        "title": "image",
                        "description": "image",
                        "type": "string",
                        "format": "string",
                        "example": "image_url"
                    },
                    "inventory_stock": {
                        "title": "inventory_stock",
                        "description": "inventory_stock",
                        "type": "integer",
                        "format": "string",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "SwaggerProductResource": {
                "title": "ProductResource",
                "description": "Product resource",
                "xml": {
                    "name": "SwaggerProductResource"
                },
                "allOf": [
                    {
                        "$ref": "#/components/schemas/PaginatedResource"
                    },
                    {
                        "properties": {
                            "payload": {
                                "title": "Payload",
                                "description": "Payload wrapper",
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/SwaggerProduct"
                                }
                            }
                        }
                    }
                ]
            },
...
            "PaginatedResource": {
                "title": "PaginatedResource",
                "description": "PaginatedResource model",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "status",
                        "type": "string",
                        "format": "string",
                        "example": "ok"
                    },
                    "message": {
                        "title": "message",
                        "description": "message",
                        "type": "string",
                        "format": "string",
                        "example": "Items listed successfully"
                    },
                    "meta_data": {
                        "title": "Payload",
                        "description": "Payload wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MetaData"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PaginatedResource"
                }
            }
        }
    }
}